django-aws/requirements.txt
asgiref==3.7.2
boto3==1.33.4
botocore==1.33.4
colorama==0.4.6
Django==4.2.7
django-crispy-forms==1.14.0
django-htmx==1.17.2
django-storages==1.14.2
factory-boy==3.3.0
Faker==21.0.0
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
pytest==7.4.3
pytest-django==4.7.0
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
s3transfer==0.8.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
